Completely dismantle GitMacheteGraphTableManager
