--- frontendUiApi/src/main/java/com/virtuslab/gitmachete/frontend/ui/api/table/IGraphTableManager.java
@@ -1,5 +1,0 @@
-package com.virtuslab.gitmachete.frontend.ui.api.table;
-
-public interface IGraphTableManager {
-  void queueRepositoryUpdateAndGraphTableRefresh();
-}
