--- frontendUiApi/src/main/java/com/virtuslab/gitmachete/frontend/ui/api/table/IGraphTableManagerFactory.java
@@ -1,10 +1,0 @@
-package com.virtuslab.gitmachete.frontend.ui.api.table;
-
-import com.intellij.openapi.project.Project;
-
-import com.virtuslab.gitmachete.frontend.ui.api.root.IGitRepositorySelectionProvider;
-
-public interface IGraphTableManagerFactory {
-  IGraphTableManager create(BaseGraphTable graphTable, Project project,
-      IGitRepositorySelectionProvider gitRepositorySelectionProvider);
-}
